split subscribe and footer
